requirements.txt
bcrypt==4.2.1
fastapi==0.115.8
google-generativeai
numpy==2.2.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
pydantic==2.10.6
python-jose==3.3.0
requests==2.32.3
scikit-learn==1.6.1
SQLAlchemy==2.0.37
tqdm==4.67.1
uvicorn==0.34.0
python-multipart